BUGFIX: Write 0 thunk rva if arch is 64 bit.

+++ src/AsmResolver.PE/Imports/Builder/ThunkTableBuffer.cs
@@ -82,7 +82,7 @@ namespace AsmResolver.PE.Imports.Builder
             if (Is32Bit)
                 writer.WriteUInt32(0);
             else
-                writer.WriteUInt16(0);
+                writer.WriteUInt64(0);
         }
         
     }
